requirements.txt
PyPDF2
pandas
numpy
langchain-groq
langgraph